refactor: renomeando pasta src/fast_api => src/api e arrumando referencias externas
