+++ main.py
@@ -21,7 +21,7 @@ if(flags.etl == True or flags.etl_only == True):
 
 # --- API ---
 
-api_root_path = "src.fast_api.app"
+api_root_path = "src.api.app"
 port = 8000
 config = uvicorn.Config(api_root_path + ":app", port=port, log_level="info", reload=True)
 server = uvicorn.Server(config)
