+++ src/api/endpoints/routes.py
@@ -1,11 +1,10 @@
 from fastapi import APIRouter, HTTPException
-from src.fast_api.models import Question, Response
-from src.fast_api.controller.AskController import AskController
+from src.api.models import Question, Response
+from src.api.controller.AskController import AskController
+
 import os 
 from dotenv import load_dotenv
-
 load_dotenv() # Loads env variables
-
 GEMINI_MODEL_NAME = os.getenv("GEMINI_MODEL_NAME")
 
 router = APIRouter()
